requirements/base.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
drf-spectacular>=0.27,<1.0
Pillow>=10.4,<11.0
requests>=2.32,<3.0
psycopg[binary]>=3.2,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
